pyproject.toml
[project]
name = "xai-cookbook"
version = "0.1.0"
description = "Comprehensive xAI API cookbook with examples for all endpoints and models"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.40.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.8.0",
    "types-requests>=2.32.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
